requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.9
importlib-metadata==1.6.0
lxml==4.5.0
more-itertools==8.2.0
multidict==4.7.5
packaging==20.3
parse==1.15.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyee==7.0.2
pymongo==3.10.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pytest==5.4.2
requests==2.23.0
requests-html==0.10.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tqdm==4.46.0
urllib3==1.25.9
w3lib==1.22.0
wcwidth==0.1.9
webdriver-manager==2.5.0
websockets==8.1
yarl==1.4.2
zipp==3.1.0
